blocking backend routes based on email

+++ backend/app.py
@@ -4,15 +4,42 @@ from flask_cors import CORS
 from google_client.GoogleClient import GoogleClient
 app = Flask(__name__)
 CORS(app)
+
+def valid_token(access_token, refresh_token):
+    google = GoogleClient(access_token=access_token, refresh_token=refresh_token)
+    try:
+        google.get_email()
+        return True
+    except:
+        return False
+
+def admin(access_token, refresh_token):
+    google = GoogleClient(access_token=access_token, refresh_token=refresh_token)
+    try:
+        email= google.get_email()
+        return email == "keeganasmith2003@gmail.com"
+    except:
+        return False
+
 @app.route('/get_posts')
 def get_posts():
-    client = Database_Client()
-    my_posts= client.get_posts()
-    return jsonify(my_posts)
+    access_token = request.args["access_token"]
+    refresh_token = request.args["refresh_token"]
+    if(valid_token(access_token, refresh_token)):
+        client = Database_Client()
+        my_posts= client.get_posts()
+        return jsonify(my_posts)
+    else:
+        return jsonify({"status": "error", "message": "Token invalid"}), 401
+
 @app.route('/new_post', methods = ['POST'])
 def new_post():
-    client = Database_Client()
     data = request.get_json()
+    access_token = data['access_token']
+    refresh_token = data['refresh_token']
+    if(not admin(access_token, refresh_token)):
+        return jsonify({"status": "error", "message": "Token invalid"}), 401
+    client = Database_Client()
     response = client.new_post(data)
     if response['ResponseMetadata']['HTTPStatusCode'] == 200:
         return jsonify({"status": "success", "message": "Post created successfully!"}), 200
@@ -21,8 +48,12 @@ def new_post():
     
 @app.route('/delete_post', methods= ['POST'])
 def del_post():
-    client = Database_Client()
     data = request.get_json()
+    access_token = data['access_token']
+    refresh_token = data['refresh_token']
+    if(not admin(access_token, refresh_token)):
+        return jsonify({"status": "error", "message": "Token invalid"}), 401
+    client = Database_Client()
     response = client.delete_item(p_key=data["p_key"], s_key=data["s_key"])
     if response['ResponseMetadata']['HTTPStatusCode'] == 200:
         return jsonify({"status": "success", "message": "Post deleted successfully!"}), 200
@@ -31,7 +62,6 @@ def del_post():
 
 @app.route('/get_email', methods= ['GET'])
 def get_email():
-    print(request.args)
     access_token = request.args["access_token"]
     refresh_token = request.args["refresh_token"]
     google = GoogleClient(access_token=access_token, refresh_token=refresh_token)
@@ -40,5 +70,6 @@ def get_email():
         return jsonify({"status": "success", "message": my_email}), 200
     except:
         return jsonify({"status": "error", "message": "Token invalid"}), 401
+
 if __name__ == '__main__':
     app.run(debug=True, host="0.0.0.0")
